choir: updating bcm structure for unit tests
