+++ src/common_includes/can_id_list.h
@@ -1,6 +1,7 @@
 #ifndef CAN_ID_LIST_H
 #define CAN_ID_LIST_H
 
-#define CAN_ID_COMMAND   (0x111U)
+#define CAN_ID_SENSOR_READ   (0x110U)
+#define CAN_ID_COMMAND       (0x111U)
 
 #endif
